MAINT: extract displacement calculations
Relevant methods are extracted from `FloeCoupled` to loosen their
coupling to instantiated objects. This greatly easens testing, as the
relevant parameters can be input directly rather than having to be
determined from "more raw" parameters, for example wavenumbers do not
have to be computed from a period and a dispersion relation, etc.