ENH: add possibility to grow waves
Add two fields to `Domain` objects to parametrise a semi-Gaussian kernel
used for wave growth. Waves to the left of the mean parameter are fully
developped, while waves to the right see their amplitudes reduced in
relation to the standard deviation parameter. These parameters are
optional.
The treatment of these growing waves (BVP, elastic energy, etc.) is done
entirelly numerically. This commit adds the possibility to treat
numerically the case of developped waves, that was entirelly treated
analytically until now.