ENH: introduce Experiment
We introduce an `Experiment` class, whose purpose is to hold the history
of a floe array. Instead of only storing the current floes in a `Domain`
object, a `Domain` object is associated to an `Experiment` object (and
still handle the logic of coupling fluid, waves and ices by solving
dispersion relations) which keeps track of timesteps and the state of
the array of floes at each of these timesteps.