MAINT(ice): refactor IceCoupled, drop pydantic
Previous commits introduced validation through `pydantic`, which is
first a serialisation library, and was probably overkill for the
intended goal: reduce manual boilerplate when adding new classes. This
commit removes `pydantic` as a dependecy, and delegates class decoration
to `attrs`.
Additionally, the IceCoupled class is removed and replaced with:
* a FloatingIce class, that subclasses Ice and adds the attributes
  related to the fact that Ice floats on Ocean, such as draft: it does
  not need any wave-related information;
* a WaveUnderIce class, which composes with FloatingIce, and add a
  `wavenumbers` attribute as well as the derived properties such as
  `attenuations`.
`attrs`-helped classes whose parametres depend on complex objects (such
as WaveUnderIce depending on FloatingIce) are initialised through class
method factories. In particular, in the case of WaveUnderIce, a new
class is added whose sole purpose is to determine the wavenumbers. This
should facilitate handling different dispersion relations in later
releases.

MAINT: rename IceFloating to FloatingIce