MAINT(refac): restructure the package
Most of the model code sat in a single file at the root of the package.
This commit splits it into two subpackages model/ and api/.
The former contains most of what can be thought of physical objects (for
example, Ice), as well as the classes responsible for the fracture
logic. The latter contains classes that should be exposed and acted upon
by a user.
The goal was to eliminate unhealthy circular imports.