MAINT(refac): remove code duplicates
WavesUnderFloe objects had a Floe attribute, with itself an Ice
attribute, as well as a WavesUnderIce attribute, with itself a
FloatingIce attribute, extending the Floe's Ice attribute.
WavesUnderFloe and Floe defined ordering methods in the same way. This
commits introduces a Subdomain parent class, extended by both Floe and
WavesUnderFloe, and removes the Floe attribute of WavesUnderFloe
instances, to eliminate these redundancies.
Additionally, the commit adds two methods to WavesUnderFloe:
* copy() returns a shallow copy of self, at the exception of the mutable
  field edge_amplitudes, which is deep-copied to prevent side effects;
* shift() allows for changing edge_amplitudes in place, in order to not
  have to rebuild a SortedList from scratch at each iteration. The list
  of subdomains, attribute of a Domain instance, can be updated in place
  on calling domain.breakup(), and then updated in place when calling
  domain.iterate(), before a tuple-copy of domain.subdomains is made
  with experiment.save_step().