MAINT,BUG: merge physics subfiles into main file
The functions to compute analytical displacement, curvature, and energy,
are moved to methods of their respective handlers.
The rationale was to allow for easy decoration of the analytical
versions of `compute`, while not decorating the numerical versions.
As the numerical version returns a scalar, applying the decorator led to
a TypeError when calling `len`.
Instead of duplicating the decorator in multiple files, these files were
merged.