ENH: rename FreeSurfaceHandler to FluidSurfaceHandler
This class computes the vertical distance between the fluid interface,
be it with air or a floe, and a reference (for instance, free surface at
rest or base of a floe at rest). Therefore, the name FreeSurface was not
really appropriate.
In addition, this commit adds a dedicated initialiser to instantiate a
handler aimed at representing the fluid--air interface.