ENH(wip): add support for pseudo-scattering after breakup
Physically, it is unrealistic that the right edge of the floe left of a
fracture, and the left edge of a floe right of a fracture, will stay in
phase. This is supported by simple experiments conducted with `swisip`.
To simulate scattering, without solving for it, we add classes defined
in `lib.phase_shift.py` that add some randomness to the
`edge_amplitudes` attribute of newly formed floes.