MAINT: simplify function and its calls
The function `flexfrac1d.lib.physics._egy_par_vals` received
`c_wavenumbers` as argument, and returned it unchanged alongside the
values it was computing. That returned `c_wavenumbers` were in turn used
in some of the scopes that were calling the function in the first place.
The function has been streamlined.