ENH(experiment): add interface to save and load
These objects can grow quite large if left to run for a while.
This commit adds methods to write the current object (including the
whole history) to disk, before clearing the history, therefore limiting
the footprint of the running script. The last state is then used to
initialise a new history, with the current time.
A function to return an `Experiment` from a file is also added.

For convenience, and because of the many data structures used, we rely
on pickle for now.