ENH: add pseudo analytical integration
The numerical solution to the BVP is a piecewise C1 cubic. As long as we
use the linear definition of curvature, computing energy requires
integrating the square of a cubic, that is, a 6th order polynomial. This
polynomial can be constructed from the coefficients of the cubic.
Computing its integral is then straightforward.
This commit adds the option (made default) to use this method to compute
energy, instead of relying on a numerical integrator.