MAINT: refactor dr to extract logic from model
Root finding of the dispersion relation rests on computing some summary
numbers; in the case of the open water linear DR, such a number is the
squared angular frequency divided by the water depth, usually written
alpha. The logic behind the computation of this number was a part of the
model class FreeSurfaceWaves, which had to be aware of it to instantiate
a FreeSurfaceSolver object.
This logic is extracted from FreeSurfaceWaves, and FreeSurfaceSolver
receives a class method that handles it.